+++ packages/rag/src/document/schema/index.ts
@@ -1,3 +1,3 @@
-export * from './node';
-export { TransformComponent } from './types';
-export type { BaseOutputParser } from './types';
+export { BaseNode, Document, TextNode } from './node';
+export { NodeRelationship, ObjectType } from './types';
+export type { Metadata, RelatedNodeInfo, RelatedNodeType, BaseNodeParams } from './types';
